+++ src/screens/ModerationInbox/index.tsx
@@ -1,5 +1,5 @@
 import {useState} from 'react'
-import {Pressable, View} from 'react-native'
+import {View} from 'react-native'
 import {plural} from '@lingui/core/macro'
 import {Trans, useLingui} from '@lingui/react/macro'
 
@@ -7,12 +7,10 @@ import {Pager} from '#/view/com/pager/Pager'
 import {TabBar} from '#/view/com/pager/TabBar'
 import {NotFoundScreen} from '#/view/screens/NotFound'
 import {atoms as a, useTheme} from '#/alf'
-import {ChevronBottom_Stroke2_Corner0_Rounded as ChevronDownIcon} from '#/components/icons/Chevron'
 import * as Layout from '#/components/Layout'
 import {createStaticClick, SimpleInlineLinkText} from '#/components/Link'
-import * as Menu from '#/components/Menu'
-import {Text} from '#/components/Typography'
 import {useAnalytics} from '#/analytics'
+import {FilterMenu} from './components/FilterMenu'
 import {ReportRow} from './components/ReportRow'
 
 type ReportFilter = 'all' | 'pending' | 'resolved' | 'unread'
@@ -60,26 +58,10 @@ export function ModerationInboxScreen() {
 function YourReports() {
   const t = useTheme()
   const {t: l} = useLingui()
+
   const [filter, setFilter] = useState<ReportFilter>('all')
 
   const hasUnread = true // TODO This is hard-coded atm. -dsb
-
-  const labels: Record<ReportFilter, string> = {
-    all: l({context: 'moderation-report-filter', message: 'All'}),
-    pending: l({
-      context: 'moderation-report-filter',
-      message: 'Under review',
-    }),
-    resolved: l({context: 'moderation-report-filter', message: 'Resolved'}),
-    unread: l({context: 'moderation-report-filter', message: 'Unread'}),
-  }
-  const options: {value: ReportFilter; label: string}[] = [
-    {value: 'all', label: labels.all},
-    {value: 'pending', label: labels.pending},
-    {value: 'resolved', label: labels.resolved},
-    {value: 'unread', label: labels.unread},
-  ]
-  const currentLabel = labels[filter]
 
   // TODO Placeholders. - dsb
   const account = '@deleteme01.bsky.social'
@@ -99,37 +81,7 @@ function YourReports() {
           t.atoms.border_contrast_low,
           {minHeight: 48},
         ]}>
-        <Menu.Root>
-          <Menu.Trigger
-            label={l`Filter moderation reports (currently: ${currentLabel})`}>
-            {({props}) => (
-              <Pressable
-                {...props}
-                style={[
-                  a.flex_row,
-                  a.align_center,
-                  a.justify_center,
-                  a.gap_xs,
-                ]}>
-                <Text style={[a.text_md, a.font_medium]}>{currentLabel}</Text>
-                <ChevronDownIcon size="xs" style={[t.atoms.text]} />
-              </Pressable>
-            )}
-          </Menu.Trigger>
-          <Menu.Outer>
-            <Menu.Group>
-              {options.map(option => (
-                <Menu.Item
-                  key={option.value}
-                  label={option.label}
-                  onPress={() => setFilter(option.value)}>
-                  <Menu.ItemText>{option.label}</Menu.ItemText>
-                  <Menu.ItemRadio selected={filter === option.value} />
-                </Menu.Item>
-              ))}
-            </Menu.Group>
-          </Menu.Outer>
-        </Menu.Root>
+        <FilterMenu filter={filter} setFilter={setFilter} />
         {hasUnread ? (
           <SimpleInlineLinkText
             label={l`Mark all reports as read`}
@@ -172,7 +124,12 @@ function YourReports() {
 }
 
 function YourAccount() {
+  const t = useTheme()
   const {t: l} = useLingui()
+
+  const [filter, setFilter] = useState<ReportFilter>('all')
+
+  const hasUnread = true // TODO This is hard-coded atm. -dsb
 
   // TODO Placeholders. - dsb
   const guideline = l({
@@ -190,6 +147,30 @@ function YourAccount() {
 
   return (
     <Layout.Center>
+      <View
+        style={[
+          a.flex_row,
+          a.align_center,
+          a.justify_between,
+          a.gap_lg,
+          a.px_lg,
+          a.py_sm,
+          a.border_b,
+          t.atoms.border_contrast_low,
+          {minHeight: 48},
+        ]}>
+        <FilterMenu filter={filter} setFilter={setFilter} />
+        {hasUnread ? (
+          <SimpleInlineLinkText
+            label={l`Mark all actions as read`}
+            style={[a.text_md, t.atoms.text]}
+            {...createStaticClick(() => {
+              // TODO Handle this action. -dsb
+            })}>
+            <Trans>Mark all as read</Trans>
+          </SimpleInlineLinkText>
+        ) : undefined}
+      </View>
       <ReportRow
         subject={l({
           context: 'moderation-report-action',
